% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_def.R
\docType{data}
\name{andy}
\alias{andy}
\title{andy}
\format{Adata frame of hamburger franchises in 75 cities from Big Andy's Burger Barn.
\describe{
 \item{sales}{Monthly sales revenue ($1000s)}
 \item{price}{A price index for all products sold in a given month}
 \item{advert}{Expenditure on advertising ($1000s)}
 }}
\source{
Courtesy of hamburger franchise, called Big Andy's Burger Barn
}
\usage{
andy
}
\description{
andy
}
\keyword{datasets}

