% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_def.R
\docType{data}
\name{bangla}
\alias{bangla}
\title{bangla}
\format{A data frame with 34 rows and 2 variables:
\describe{
  \item{p}{Price of sugar cane divided by the price of jute}
  \item{a}{area planted in thousands of hectares in a region of Bangladesh.}
}}
\source{
\url{http://www.principlesofeconometrics.com/poe4/poe4stata.htm}
}
\usage{
bangla
}
\description{
bangla
}
\keyword{datasets}

