% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_def.R
\docType{data}
\name{beer}
\alias{beer}
\title{beer}
\format{30 annual observations from a single household
\describe{
  \item{q}{litres of beer consumed}
  \item{pb}{Price of beer ($)}
  \item{pl}{price of other liquor ($)}
  \item{pr}{price of remaining goods and services (an index)}
  \item{i}{income ($)}
}}
\source{
\url{http://www.principlesofeconometrics.com/poe4/poe4stata.htm}
}
\usage{
beer
}
\description{
beer
}
\keyword{datasets}

