% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_def.R
\docType{data}
\name{bond}
\alias{bond}
\title{bond}
\format{data frame; obs: 102, monthly, (1968:1 - 1976.6)
\describe{
 \item{bond}{AA railroad yields}
 }}
\usage{
bond
}
\description{
bond
}
\keyword{datasets}

