% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_def.R
\docType{data}
\name{br}
\alias{br}
\title{br}
\format{Obs:   1080 home sales in Baton Rouge, LA during mid-2005
\describe{
  \item{price}{sale price, dollars}
  \item{sqft}{total square feet}
  \item{bedrooms}{number of bedrooms}
  \item{baths}{number of full baths}
  \item{age}{age in years}
  \item{Occupancy}{Owner = 1
                   Vacant = 2
                   Tenant = 3}
  \item{Pool}{Yes = 1, No = 0}
  \item{Style}{
  Traditional = 1
  Townhouse = 2
  Ranch = 3
  New Orleans = 4
  Mobile Home = 5
  Garden = 6
  French = 7
  Cottage = 8
  Contemporary = 9
  Colonial = 10
  Acadian = 11
  }
  \item{Fireplace}{Yes = 1, No = 0}
  \item{Waterfront}{Yes =1, No = 0}
  \item{DOM}{Days on the market}
}}
\source{
Dr. Kelley Pace, Department of Finance,
         Louisiana State University
}
\usage{
br
}
\description{
br
}
\keyword{datasets}

