\name{vacation}
\alias{vacation}
\docType{data}
\title{
Vacation Data
}
\description{
Obs:   200
}
\usage{data("vacation")}
\format{
  A data frame with 200 observations on the following 4 variables.
  \describe{
    \item{\code{miles}}{miles traveled per year}
    \item{\code{income}}{annual income in $1000's}
    \item{\code{age}}{average age of adult members of household}
    \item{\code{kids}}{number of children in household}
  }
}
\details{
A sample of Chicago households.
}
\source{
http://principlesofeconometrics.com/poe4/poe4.htm
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(vacation)
## maybe str(vacation) ; plot(vacation) ...
}
\keyword{datasets}
